% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOPART.R
\name{FLOPART}
\alias{FLOPART}
\title{Functional Labeled Optimal Partitioning}
\usage{
FLOPART(coverage, label, penalty)
}
\arguments{
\item{coverage}{data frame of coverage}

\item{label}{data frame of labels}

\item{penalty}{non-negative penalty constant}
}
\value{
augmented list of output from FLOPART_interface and
  FLOPART_data
}
\description{
Main function for computing optimal segmentation model with
Poisson loss, up-down constraints, and label constraints.
}
\examples{
data("Mono27ac", package="FLOPART")
fit <- with(Mono27ac, FLOPART::FLOPART(coverage, labels, 1e6))
str(fit)
}
\author{
Toby Dylan Hocking
}
