% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FLOPART_interface}
\alias{FLOPART_interface}
\title{Interface to FLOPART C++ code}
\usage{
FLOPART_interface(
  data_vec,
  weight_vec,
  penalty,
  label_type_vec,
  label_start_vec,
  label_end_vec
)
}
\arguments{
\item{data_vec}{Integer vector of non-negative count data}

\item{weight_vec}{Numeric vector of positive weights (same size as data_vec)}

\item{penalty}{non-negative real-valued penalty (larger for fewer peaks)}

\item{label_type_vec}{Integer vector of label types}

\item{label_start_vec}{Integer vector of label starts}

\item{label_end_vec}{Integer vector of label ends}
}
\value{
List with named elements: cost_mat and intervals_mat (one row for each data point, first column up, second down), segments_df (one row for each segment in the optimal model)
}
\description{
Interface to FLOPART C++ code
}
