% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_label_code}
\alias{get_label_code}
\title{Lookup the integer values used to represent different label types}
\usage{
get_label_code()
}
\value{
Integer vector with names corresponding to supported label types
}
\description{
Lookup the integer values used to represent different label types
}
