\name{Mono27ac}
\alias{Mono27ac}
\docType{data}
\title{
  H3K27ac ChIP-seq data from one Monocyte sample
}
\description{
  Raw coverage data and labels are provided in order to test the FLOPART algo.
}
\usage{data("Mono27ac")}
\format{
List of 2 data.tables, coverage and labels.
}
\source{
https://github.com/tdhock/feature-learning-benchmark
}
\keyword{datasets}
