\name{Mono27ac.simple}
\alias{Mono27ac.simple}
\docType{data}
\title{
  Smaller H3K27ac ChIP-seq data from one Monocyte sample
}
\description{
  Raw coverage data and labels are provided in order to test the FLOPART algo.
}
\usage{data("Mono27ac.simple")}
\format{
  List of two data tables, coverage and label.
}
\source{
https://github.com/tdhock/feature-learning-benchmark
}
\keyword{datasets}
